Re-worked merging algorithm, using Github 'commit' API.
The previous implementation, which was based on Github 'merge' API,
has a flaw: after merge into auto_branch is done, there is no way
to undo it. It makes no problem if auto_branch tests are successful,
but if they are not, we end up with closed PR and inability to
finish our merge into master.

There is no ideal solution for this problem, so our trade of is
the following: do not use 'merge' button API and do manual merge
instead. The disadvantage is after merging the PR gets only
'closed' status (instead of 'merged' and 'closed'), confusing the
reader that the PR became abandoned for some reason. To eliminate
this problem we decided to mark the PR with a label(not done yet).

This solution makes use of Github 'merge' revision, stored as
a reference under 'refs/pull/:pr/merge' reference path, which
is created by Github for each PR. With Github 'commit' API,
we create a new commit, pointing to this 'merge' revision and reset
auto_branch into it. If checks are successful, we merge master
with FF mode into 'auto', closing the PR.