Restore merging context on startup.
Now, when bot starts, it checks whether there is a still merging
auto_branch revision. If there is one, it completes merging process.
This is needed primary for a situation when the bot was terminated while
it was waiting for auto_branch checks completing (for a PR had been
merged into auto_branch).