Properly handle a situation when the bot was terminated during cleanup
When the PR staging commit was merged on it's base, there are still some
cleanup steps to be executed, such as PR labeling, closing and removing
tag. After this fix, the bot will try to complete these cleanup steps
even if it was killed during them(or an unexpected error occurred).