Properly handle unexpected terminations on resume merging
This is a second part of modifications, started in 348a9c0.  Before this
change, all unexpected errors during resuming PR merging caused a
general cleanup: removing the tag and adjusting labels (Merge-failed).
However such approach is not reliable and can leave the PR in an
inconsistent state.  For example, if an error occurred during
fast-forwarding, we end up with still open PR but actually merged PR.
If an accidental error occurs before fast-forwarding step(while
extracting information about 'current' PR), we can't remove the
tag before we are sure that the tag revision is not in master already
(in a way without bot participation). Also, it looks better to repeat
merge attempt (after 10 minutes) instead of immediately removing the tag
and discarding valuable staging test results(which could last for
hours).

This change addresses a TODO from 348a9c0: compare commits in order
to properly check whether a revision has been merged.